+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `laravel-livewire-tables` will be documented in this file
 
+## UNRELEASED
+### New Features
+- Add capability to customise colors/styling on the Pagination Dropdown by @lrljoe
+
 ## [v3.2.2] - 2024-02-29
 ### New Features
 - Add setDefaultPerPage by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1671